to organize dto files
