+++ back/src/main/java/com/openclassrooms/mddapi/dtos/article/CreateArticleDto.java
@@ -1,4 +1,4 @@
-package com.openclassrooms.mddapi.dtos;
+package com.openclassrooms.mddapi.dtos.article;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
