+++ back/src/main/java/com/openclassrooms/mddapi/dtos/user/RegisterUserDto.java
@@ -1,4 +1,4 @@
-package com.openclassrooms.mddapi.dtos;
+package com.openclassrooms.mddapi.dtos.user;
 
 import lombok.Data;
