+++ back/src/main/java/com/openclassrooms/mddapi/services/ArticleService.java
@@ -1,6 +1,6 @@
 package com.openclassrooms.mddapi.services;
 
-import com.openclassrooms.mddapi.dtos.*;
+import com.openclassrooms.mddapi.dtos.article.*;
 import com.openclassrooms.mddapi.models.Article;
 import com.openclassrooms.mddapi.models.Theme;
 import com.openclassrooms.mddapi.models.User;
